file-analyzer/requirements.txt
fastapi
uvicorn
jinja2
PyPDF2
python-docx